Polices added

+++ app/Http/Controllers/Post/DestroyController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Post;
 
 use App\Models\Post;
 use App\Http\Controllers\Controller;
+use Illuminate\Http\Request;
 
 class DestroyController extends Controller
 {
@@ -13,8 +14,12 @@ class DestroyController extends Controller
    * @param  \App\Models\Post  $post
    * @return \Illuminate\Http\Response
    */
-  public function __invoke(Post $post)
+  public function __invoke(Request $request, Post $post)
   {
+    if ($request->user()->cannot('delete', $post)) {
+      abort(403);
+    }
+
     if ($post->image) unlink($post->image);
     $post->destroy($post->id);
     return redirect()->route('posts.index');
